Only set crossorigin image prop when using canvas
The crossorigin image property is mainly useful when the image is loaded
into a canvas element, to prevent the canvas from becoming tainted and
thus impossible to extract data from.

Setting the crossorigin property on an image will trigger CORS policies
when loading images, and will prevent images from being loaded if
they're hosted on a cross origin without appropriate CORS headers set in
the response. Being able to crop externally hosted images is still
useful even if the external host does not set CORS headers, with the
only limitation of not being able to rotate the image.