Implement initial controls for resizing viewport
An option to enable resizing has to be set to `true`, while further
customization of which axis can be resized can also be configured.

With resizing enabled a resize handle at the bottom and/or the right of
the viewport will be displayed. Hovering over it will change the cursor
to indicate resizability. Dragging this handle will update the viewport
and boundary to the new size interactively, but the axis will be
constrained to the end of the image.