Bound max resize width/height to full image width/height
We previously bounded the maximum height or width one could resize the
image up to according to how much of the image was left below or to the
right of the viewport. But after integrating more deeply with croppie's
other updates we don't need to do that. We can simply limit it to the
actual image height or width and croppie will ensure to center the image
appropriately.