Support resize of viewport (#347)
* Implement initial controls for resizing viewport

An option to enable resizing has to be set to `true`, while further
customization of which axis can be resized can also be configured.

With resizing enabled a resize handle at the bottom and/or the right of
the viewport will be displayed. Hovering over it will change the cursor
to indicate resizability. Dragging this handle will update the viewport
and boundary to the new size interactively, but the axis will be
constrained to the end of the image.

* Trigger update event on each resize update

* Enforce a minimum size during resize and rename option

* Keep min/max zoom values in sync with resized viewport

* Assume 1.0 as scale when calculating minimum zoom initially

* Bound max resize width/height to full image width/height

We previously bounded the maximum height or width one could resize the
image up to according to how much of the image was left below or to the
right of the viewport. But after integrating more deeply with croppie's
other updates we don't need to do that. We can simply limit it to the
actual image height or width and croppie will ensure to center the image
appropriately.

* Some code cleanup

* Add default styling of resize controls

* Add demo of resizer controls