Fix reading EXIF orientation when 'enableOrientation' is true (#499)
EXIF orientation data got ignored because of commit 018d30c
(which fixed displaying image when exif-js is not available).

This only happens when 'enableOrientation' is true and you do
not pass an orientation to 'bind'. This is for example the case
when you want to add a button to rotate the image.

You can reproduce this issue in the demo 'upload-demo' when you
set 'enableOrientation' to true.