Don't silently discard image loading errors
The `bind` method returns a promise which mainly awaits loading the
image URL. If an error happened within the image loader, then the `bind`
method would simply log it to the console and the caller's promise
handling would continue as if everything was fine. In addition, network
issues or otherwise errors in the browser's loading of the image was not
caught and propagated back up to the `bind` method.

This implements a simple `onerror` handler for loading the image which
rejects the loading promise. The `catch` handler in the `bind` method
which swallowed any rejections have also been removed, which enables the
caller to handle this properly on their side.